# internal/server/server.go
package server

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/rs/zerolog"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
	"gorm.io/gorm"

	"learning-go-shop/internal/config"
	"learning-go-shop/internal/services"
)

type Server struct {
	config         *config.Config
	db             *gorm.DB
	logger         *zerolog.Logger
	authService    *services.AuthService
	productService *services.ProductService
	userService    *services.UserService
	uploadService  *services.UploadService
	cartService    *services.CartService
	orderService   *services.OrderService
}

func New(cfg *config.Config,
	db *gorm.DB,
	logger *zerolog.Logger,
	authService *services.AuthService,
	productService *services.ProductService,
	userService *services.UserService,
	uploadService *services.UploadService,
	cartService *services.CartService,
	orderService *services.OrderService,

) *Server {
	return &Server{
		config:         cfg,
		db:             db,
		logger:         logger,
		authService:    authService,
		productService: productService,
		userService:    userService,
		uploadService:  uploadService,
		cartService:    cartService,
		orderService:   orderService,
	}
}

func (s *Server) SetupRoutes() *gin.Engine {
	router := gin.New()
	router.Use(gin.Logger())
	router.Use(gin.Recovery())
	router.Use(s.corsMiddleware())
	router.GET("/health", s.healthCheck)
	router.Static("/uploads", "./uploads")

	// Add documentation routes
	router.GET("/docs/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	router.StaticFile("/api-docs", "./docs/rapidoc.html")

	router.Static("/uploads", "./uploads")

	api := router.Group("/api/v1")
	{

		auth := api.Group("/auth")
		{
			auth.POST("/register", s.register)
			auth.POST("/login", s.login)
			auth.POST("/refresh", s.refreshToken)
			// auth.POST("/logout", s.register)1

			{
				protected := auth.Group("/")
				protected.Use(s.authMiddleware())
				{
					users := protected.Group("users")
					{
						userRoutes := users
						userRoutes.GET("/profile", s.getProfile)
						userRoutes.POST("/profile", s.updateProfile)
					}

				}
				// category routes
				categories := protected.Group("/categories")
				{
					categoryRoute := categories
					categoryRoute.POST("/", s.adminMiddleware(), s.createCategory)
					categoryRoute.PUT("/:id", s.adminMiddleware(), s.updateCategory)
					categoryRoute.DELETE("/:id", s.adminMiddleware(), s.deleteCategory)
				}

				// product routes
				products := protected.Group("/products")
				{
					productRoutes := products
					productRoutes.POST("/", s.adminMiddleware(), s.createProduct)
					productRoutes.PUT("/:id", s.adminMiddleware(), s.updateProduct)
					productRoutes.DELETE("/:id", s.adminMiddleware(), s.deleteProduct)
					productRoutes.POST("/:id/images", s.adminMiddleware(), s.uploadProductImage)

				}
				// cart routes
				cart := protected.Group("/cart")
				{
					cartRoutes := cart
					cartRoutes.GET("/", s.getCart)
					cartRoutes.POST("/items", s.addToCart)
					cartRoutes.PUT("/items/:id", s.updateCartItem)
					cartRoutes.DELETE("/items/:id", s.removeFromCart)
				}
				orders := protected.Group("/orders")
				{
					orderRoutes := orders
					orderRoutes.POST("/", s.createOrder)
					orderRoutes.GET("/", s.getOrders)
					orderRoutes.GET("/:id", s.getOrder)
				}
			}

		}
		api.GET("/categories", s.getCategories)
		api.GET("/search", s.searchProducts)
		api.GET("/products", s.getProducts)
		api.GET("/products/:id", s.getProduct)
	}
	// asdsdasd
	return router
}

func (s *Server) healthCheck(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"status": "ok"})
}

func (s *Server) corsMiddleware() gin.HandlerFunc {
	return func(ctx *gin.Context) {
		ctx.Header("Access-Control-Allow-Origin", "*")
		ctx.Header("Access-Control-Allow-Method", "GET,POST,PUT,DELETE,OPTIONS")
		ctx.Header("Access-Control-Allow-Headers", "Context-Type,Authorization")
		if ctx.Request.Method == "QPTIONS" {
			ctx.AbortWithStatus(204)
		}
	}
}